Reorganize files by splitting them
