+++ package.json
@@ -2,8 +2,8 @@
   "name": "redux-saga-testable",
   "version": "0.3.0-alpha.1",
   "description": "Practical unit test library for redux-saga",
-  "main": "lib/runner.js",
-  "types": "lib/runner.d.ts",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
   "files": [
     "/lib"
   ],
